requirements.txt
fastapi==0.109.1
uvicorn==0.22.0
feedparser==6.0.10
pydantic==1.10.7
httpx==0.24.1
python-dotenv~=1.0.1
beautifulsoup4~=4.12.3
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
alembic==1.13.1
python-dateutil==2.8.2
openai>=1.30.0,<2